+++ lib/pay/asaas/webhooks/payment_sync.rb
@@ -4,8 +4,6 @@ module Pay
       class PaymentSync
         def call(event)
           Pay::Asaas::Charge.sync(event["payment"]["id"], object: event["payment"])
-
-          # TODO: Add notifications
         end
       end
     end