+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.1.1] - 2025-06-13
+
+- Fixes payment sync in webhooks
+- Does not create charge the not exists inside the webhook processor
+
 ## [0.1.0] - 2024-12-26
 
 - Initial release